implemented similarity detection in learning resources
