+++ controllers/serverless.txt
@@ -1,0 +1,14 @@
+Serverless architectures are application designs that incorporate third-party “Backend as a Service” (BaaS) services, and/or that include custom code run in managed, ephemeral containers on a “Functions as a Service” (FaaS) platform. By using these ideas, and related ones like single-page applications, such architectures remove much of the need for a traditional always-on server component.Serverless architectures may benefit from significantly reduced operational cost, complexity, and engineering lead time, at a cost of increased reliance on vendor dependencies and comparatively immature supporting services. Like many trends in software, there’s no one clear view of what Serverless is. 
+
+For starters, it encompasses two different but overlapping areas:
+Serverless was first used to describe applications that significantly or fully incorporate third-party, cloud-hosted applications and services, to manage server-side logic and state. These are typically “rich client” applications—think single-page web apps, or mobile apps—that use the vast ecosystem of cloud-accessible databases (e.g., Parse, Firebase), authentication services (e.g., Auth0, AWS Cognito), and so on. These types of services have been previously described as “(Mobile) Backend as a Service", and I use "BaaS" as shorthand in the rest of this article.
+Serverless can also mean applications where server-side logic is still written by the application developer, but, unlike traditional architectures, it’s run in stateless compute containers that are event-triggered, ephemeral (may only last for one invocation), and fully managed by a third party. (Note: The original source for this name—a tweet by @marak—is no longer publicly available.) AWS Lambda is one of the most popular implementations of a Functions-as-a-Service platform at present, but there are many others, too. Knative provides fundamental building blocks for serverless workloads in Kubernetes, empowering the creation of modern, container-based and cloud-native applications which can be deployed anywhere on Kubernetes. Knative is a way to abstract the operational overhead of deploying and managing workloads that run on K8s and provides a consistent approach so that developers can focus on writing cool code.
+At the time of writing there are 3 components available that can be used in addressing the tasks listed above
+Build — Source-to-container build orchestration
+Eventing — Management and delivery of events
+Serving — Request-driven compute that can scale to zero. Functions-as-a-Service (FaaS) is an event-driven computing execution model that runs in stateless containers and those functions manage server-side logic and state through the use of services. 
+
+Serverless is the architectural pattern that describe applications that combine FaaS and those hosted (managed) services.
+
+By providing the same building blocks across multiple vendors, Knative addresses the issue of portability and interoperability across different serverless platforms, which can be tremendously valuable because it allows parties building FaaS platforms to focus on providing business value as opposed to building the fundamentals themselves, while leveraging the talent and resources of a wider community.
+
